[bw] cpp typecasts and compat

+++ erts/emulator/drivers/common/ram_file_drv.cpp
@@ -119,7 +119,7 @@ struct erl_drv_entry ram_file_driver_entry = {
     NULL, /* flush */
     NULL, /* call */
     NULL, /* event */
-    ERL_DRV_EXTENDED_MARKER,
+    (int)ERL_DRV_EXTENDED_MARKER,
     ERL_DRV_EXTENDED_MAJOR_VERSION,
     ERL_DRV_EXTENDED_MINOR_VERSION,
     0,
